+++ middleman-core/lib/middleman-core/extensions/asset_hash.rb
@@ -1,5 +1,4 @@
 require 'middleman-core/util'
-require 'middleman-core/rack'
 
 class Middleman::Extensions::AssetHash < ::Middleman::Extension
   option :sources, %w[.css .htm .html .js .php .xhtml], 'List of extensions that are searched for hashable assets.'
@@ -11,9 +10,7 @@ class Middleman::Extensions::AssetHash < ::Middleman::Extension
   def initialize(app, options_hash = {}, &block)
     super
 
-    require 'addressable/uri'
     require 'digest/sha1'
-    require 'rack/mock'
 
     # Allow specifying regexes to ignore, plus always ignore apple touch icons
     @ignore = Array(options.ignore) + [/^apple-touch-icon/]
@@ -24,14 +21,6 @@ class Middleman::Extensions::AssetHash < ::Middleman::Extension
     # to be named "favicon.ico"
     @set_of_exts = Set.new(options.exts || (app.config[:asset_extensions] - %w[.ico]))
     @set_of_sources = Set.new options.sources
-
-    app.rewrite_inline_urls id: :asset_hash,
-                            url_extensions: @set_of_exts,
-                            source_extensions: @set_of_sources,
-                            ignore: @ignore,
-                            rewrite_ignore: options.rewrite_ignore,
-                            proc: method(:rewrite_url),
-                            after: :asset_host
   end
 
   Contract String, Or[String, Pathname], Any => Maybe[String]
@@ -51,16 +40,23 @@ class Middleman::Extensions::AssetHash < ::Middleman::Extension
 
     replacement_path = "/#{asset_page.destination_path}"
     replacement_path = Pathname.new(replacement_path).relative_path_from(dirpath).to_s if relative_path
-
     replacement_path
   end
 
   # Update the main sitemap resource list
   Contract IsA['Middleman::Sitemap::ResourceListContainer'] => Any
   def manipulate_resource_list_container!(resource_list)
-    @rack_client ||= begin
-      rack_app = ::Middleman::Rack.new(app).to_app
-      ::Rack::MockRequest.new(rack_app)
+    resource_list.by_exts(@set_of_sources).each do |r|
+      next if Array(options.rewrite_ignore || []).any? do |i|
+        ::Middleman::Util.path_match(i, "/#{r.destination_path}")
+      end
+
+      r.filters << ::Middleman::InlineURLRewriter.new(:asset_hash,
+                                                      app,
+                                                      r,
+                                                      url_extensions: @set_of_exts,
+                                                      ignore: options.ignore,
+                                                      proc: method(:rewrite_url))
     end
 
     # Process resources in order: binary images and fonts, then SVG, then JS/CSS.
@@ -89,15 +85,9 @@ class Middleman::Extensions::AssetHash < ::Middleman::Extension
     digest = if resource.binary?
                ::Digest::SHA1.file(resource.source_file).hexdigest[0..7]
              else
-               # Render through the Rack interface so middleware and mounted apps get a shot
-               response = @rack_client.get(
-                 ::URI.escape(resource.destination_path),
-                 'bypass_inline_url_rewriter_asset_hash' => 'true'
-               )
-
-               raise "#{resource.path} should be in the sitemap!" unless response.status == 200
-
-               ::Digest::SHA1.hexdigest(response.body)[0..7]
+               # Render without asset hash
+               body = resource.render { |f| !f.respond_to?(:filter_name) || f.filter_name != :asset_hash }
+               ::Digest::SHA1.hexdigest(body)[0..7]
              end
 
     resource_list.update!(resource, :destination_path) do