+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 master
 ===
 
+* Remove Rack support in favor of `resource.filters << proc { |oldbody| newbody }`
 * `manipulate_resource_list_container!` as a faster, less functional approach.
 
 # 4.3.0.rc.4